% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{budget_eq_prop_cte_theo}
\alias{budget_eq_prop_cte_theo}
\title{Budget initial pour porportion constante investie dans l'actif risqué}
\usage{
budget_eq_prop_cte_theo(r_s_r, cf, si, m_T, garantie, propo)
}
\arguments{
\item{r_s_r}{Taux de l'actif sans risque}

\item{cf}{Frais appliqué sur le fonds distinct}

\item{si}{Volatilité de l'actif risqué}

\item{m_T}{Maturité du fonds distinct}

\item{garantie}{Valeur de la garantie du fonds distinct}

\item{propo}{Proportion constante investie dans l'actif risqué}
}
\value{
Valeur équitable a investir dans le fonds
}
\description{
Budget initial pour porportion constante investie dans l'actif risqué
}
\examples{
budget_eq_prop_cte_theo(r_s_r=0.02,cf=0.01224,si=0.2,m_T=10,bud=1,garantie=1,propo=1)
}
