% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CRRA}
\alias{CRRA}
\title{CRRA (Constant Relative Risk Aversion)}
\usage{
CRRA(x, gamma)
}
\arguments{
\item{x}{Risky asset value}

\item{gamma}{Paramater Gamma in the Power Utility function. Must be different than 1.}
}
\value{
Utility of a CRRA utility function
}
\description{
CRRA (Constant Relative Risk Aversion)
}
\examples{
CRRA(1.5,4)
}
