% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Derivee_P_Utility_con}
\alias{Derivee_P_Utility_con}
\title{Derivative of the Concavified Power Utility function}
\usage{
Derivee_P_Utility_con(X_T_con, gamma_c, b_c, a_c, K_c)
}
\arguments{
\item{X_T_con}{Risky asset value at time t}

\item{gamma_c}{Paramater Gamma in the Power Utility function. Must be different than 0.}

\item{b_c}{The strike price of the call option (variable annuity)}

\item{a_c}{The multiplicator factor of the call option (variable annuity)}

\item{K_c}{The constant that is added to the payoff (equal to b_o most of the time)}
}
\value{
The derivative of the function Power Utility from which concavification has been made
}
\description{
Derivative of the Concavified Power Utility function
}
\examples{
Derivee_P_Utility_con(X_T_con=1.4,gamma_c=2,b_c=1,a_c=1,K_c=1)
}
