% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{derivee_Power_Utility}
\alias{derivee_Power_Utility}
\title{Derivative of the Concavified Power Utility function}
\usage{
derivee_Power_Utility(x, gamma)
}
\arguments{
\item{x}{Risky asset value}

\item{gamma}{Paramater Gamma in the Power Utility function. Must be different than 1.}
}
\value{
Derivative of Power Utility evaluated at x.
}
\description{
Derivative of the Concavified Power Utility function
}
\examples{
derivee_Power_Utility(1.5,4)
}
