% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Droite}
\alias{Droite}
\title{Droite}
\usage{
Droite(pente, origine, val_x)
}
\arguments{
\item{pente}{Pente de la droite à tracer}

\item{origine}{Origine de la droite à tracer}

\item{val_x}{Valeur de l'axe des x}
}
\value{
Valeur sur l'axe des ordonnées correspondate
}
\description{
Droite
}
\examples{
Droite(-3, 0.2,40)
}
