% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Find_lambda}
\alias{Find_lambda}
\title{Intermediate function}
\usage{
Find_lambda(
  r_FL,
  alpha_FL,
  sigma_FL,
  gamma_FL,
  K_FL,
  b_FL,
  a_FL,
  T_FL,
  lambda_FL
)
}
\arguments{
\item{r_FL}{Risk neural rate (striclty less than alpha_FL ), yearly composed}

\item{alpha_FL}{Risky rate of return, yearly composed}

\item{sigma_FL}{Volatility of the risky asset, yearly composed}

\item{gamma_FL}{Paramater Gamma in the Power Utility function. Must be different than 0}

\item{K_FL}{The constant that is added to the payoff (equal to b_o most of the time)}

\item{b_FL}{The strike price of the call option (variable annuity)}

\item{a_FL}{a_o The multiplicator factor of the call option (variable annuity)}

\item{T_FL}{Maturity, yearly composed}

\item{lambda_FL}{Lambda (Lagrange multiplicator)}
}
\value{
The initial asset value given a specific value of Lagrange multiplicator
}
\description{
This function is an intermediate function and is needed for the function Find_lambda
}
\examples{
Find_lambda(r_FL=0.02,alpha_FL=0.04,sigma_FL=0.3,gamma_FL=2,K_FL=1,b_FL=1,a_FL=1,T_FL=10,lambda_FL=0.18)
}
