% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Find_x_theta_PU}
\alias{Find_x_theta_PU}
\title{Concavification Technique}
\usage{
Find_x_theta_PU(b_o, a_o, K_o, gamma_o)
}
\arguments{
\item{b_o}{The strike price of the call option (variable annuity)}

\item{a_o}{The multiplicator factor of the call option (variable annuity)}

\item{K_o}{The constant that is added to the payoff (equal to b_o most of the time)}

\item{gamma_o}{Paramater Gamma in the Power Utility function. Must be different than 0}
}
\value{
The concavification point
}
\description{
Finds the value of the concavification point in Carpenter (2000); the point that makes an optimisation function concave again
}
\examples{
Find_x_theta_PU(1,1,1,2)
}
