% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Find_x_theta_PU_Numerous}
\alias{Find_x_theta_PU_Numerous}
\title{Concavification Technique for many insured}
\usage{
Find_x_theta_PU_Numerous(b_o, a_o, K_o, vector_gamma)
}
\arguments{
\item{b_o}{The strike price of the call option (variable annuity)}

\item{a_o}{The multiplicator factor of the call option (variable annuity)}

\item{K_o}{The constant that is added to the payoff (equal to b_o most of the time)}

\item{gamma_o}{. Vector of all the Gamma of each of the insured. Each gamma must be different than 0}
}
\value{
The concavification point of the optimization function
}
\description{
Finds the value of the concavification point in Carpenter (2000); the point that makes an optimisation function concave again for a insured people group optimisation
}
\examples{
Find_x_theta_PU_Numerous(b_o=1,a_o=1,K_o=1,vector_gamma=c(2,3,5))
}
