% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{frais_eq_prop_cte_theo}
\alias{frais_eq_prop_cte_theo}
\title{Frais équitable proportion constante: sans simulation}
\usage{
frais_eq_prop_cte_theo(r_s_r, si, m_T, bud, garantie, propo)
}
\arguments{
\item{r_s_r}{Taux de l'actif sans risque}

\item{si}{Volatilité de l'actif risqué}

\item{m_T}{Maturité du fonds distinct}

\item{bud}{Valeur initiale investie dans le fonds distinct}

\item{garantie}{Valeur de la garantie du fonds distinct}

\item{propo}{Proportion constante investie dans l'actif risqué}
}
\value{
cf: Frais équitable appliqué au fonds distinct
}
\description{
Frais équitable proportion constante: sans simulation
}
\examples{
frais_eq_prop_cte_theo(r_s_r=0.02,si=0.2,m_T=10,bud=1,garantie=1,propo=1)
}
