% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Inverse_P_Utility}
\alias{Inverse_P_Utility}
\title{Inverse Function}
\usage{
Inverse_P_Utility(uty, gamma_p)
}
\arguments{
\item{uty}{Utility value}

\item{gamma_p}{Paramater Gamma in the Power Utility function. Must be different than 0}
}
\value{
The inverse of the Power Utility function, i.e. the value of the asset that results to the corresponding utility.
}
\description{
Inverse Function
}
\examples{
Inverse_P_Utility(-0.4,2)
}
