% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Investment_Fonct}
\alias{Investment_Fonct}
\title{Rebalancing & Investment}
\usage{
Investment_Fonct(S_t, B_t, invest_risq_tm1, invest_srisq_tm1, propor)
}
\arguments{
\item{S_t}{Risky asset value at time t (time of rebalancing)}

\item{B_t}{Riskless asset value at time t (time of rebalancing)}

\item{invest_risq_tm1}{Number of units invested in the risky asset before rebalancing (at time t-1)}

\item{invest_srisq_tm1}{Number of units invested in the riskless asset before rebalancing (at time t-1)}

\item{propor}{Proportion of the total investment in the risky asset}
}
\value{
A vector which contains in first positionthe number of units invest in the risky asset, and in second the riskless
}
\description{
Rebalancing & Investment
}
\examples{
Investment_Fonct(1.4,1.2,0.3,0.7,0.5)
}
