% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lambda_optimal}
\alias{lambda_optimal}
\title{Lambda Optimal}
\usage{
lambda_optimal(
  r_FLo,
  alpha_FLo,
  sigma_FLo,
  gamma_FLo,
  K_FLo,
  b_FLo,
  a_FLo,
  X_0_FLo,
  T_FLo
)
}
\arguments{
\item{r_FLo}{Risk neural rate (striclty less than alpha_FL ), yearly composed}

\item{alpha_FLo}{Risky rate of return, yearly composed}

\item{sigma_FLo}{Volatility of the risky asset, yearly composed}

\item{gamma_FLo}{Paramater Gamma in the Power Utility function. Must be different than 0}

\item{K_FLo}{The constant that is added to the payoff (equal to b_o most of the time)}

\item{b_FLo}{The strike price of the call option (variable annuity)}

\item{a_FLo}{The multiplicator factor of the call option (variable annuity)}

\item{X_0_FLo}{Budget value, or initial value invested in the portfolio/V.A.}

\item{T_FLo}{Maturity, yearly composed}
}
\value{
The optimal Lambda (Lagrange multiplicator) given a specific intial budget value (X_0_FLo)
}
\description{
Lambda Optimal
}
\examples{
lambda_optimal(r_FLo=0.02,alpha_FLo=0.04,sigma_FLo=0.2,gamma_FLo=2,K_FLo=1,b_FLo=1,a_FLo=1,X_0_FLo=1.2,T_FLo=10)
}
