% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{P_Utility}
\alias{P_Utility}
\title{Power Utility function}
\usage{
P_Utility(X_Tu, gamma)
}
\arguments{
\item{X_Tu}{Value from which we want to calculate the utility}

\item{gamma}{Parameter of the Power utility function. Must be different than 0}
}
\value{
The (Power) utility value corresponding to inputs
}
\description{
Power Utility function
}
\examples{
P_Utility(10,2)
}
