% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{P_Utility_con}
\alias{P_Utility_con}
\title{Concavification of the Power Utility function}
\usage{
P_Utility_con(X_T_con, gamma_c, b_c, a_c, K_c)
}
\arguments{
\item{X_T_con}{Risky asset value at time t}

\item{gamma_c}{Paramater Gamma in the Power Utility function. Must be different than 0.}

\item{b_c}{The strike price of the call option (variable annuity)}

\item{a_c}{The multiplicator factor of the call option (variable annuity)}

\item{K_c}{The constant that is added to the payoff (equal to b_o most of the time)}
}
\value{
The concavified Power Utility function
}
\description{
Concavification of the Power Utility function
}
\examples{
P_Utility_con(X_T_con=1.3,gamma_c=3,b_c=1,a_c=1,K_c=1)
}
