% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{payoff_call}
\alias{payoff_call}
\title{Payoff of a call option}
\usage{
payoff_call(X_T, b, a, K)
}
\arguments{
\item{X_T}{The value of the underlying asset}

\item{b}{The strike price}

\item{a}{The multiplicator factor (1 most of the time)}

\item{K}{The constant that is added to the payoff (equal to b most of the time)}
}
\value{
The result of a call option
}
\description{
This function return the result of a call option
}
\examples{
payoff_call(10,1,1,1)
}
