% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PowerUtility_con_3}
\alias{PowerUtility_con_3}
\title{Sum of 3 concavified Power Utility functions (different risk aversion)}
\usage{
PowerUtility_con_3(gamma1, gamma2, gamma3, b, a, K, x)
}
\arguments{
\item{gamma1}{Paramater Gamma 1 in the Power Utility function. Must be different than 0.}

\item{gamma2}{Paramater Gamma 2 in the Power Utility function. Must be different than 0.}

\item{gamma3}{Paramater Gamma 3 in the Power Utility function. Must be different than 0.}

\item{b}{The strike price of the call option (variable annuity)}

\item{a}{The multiplicator factor of the call option (variable annuity)}

\item{K}{The constant that is added to the payoff (equal to b_o most of the time)}

\item{x}{Risky asset value at time t}
}
\value{
The total concavified utility coresponding to the asset value
}
\description{
Sum of 3 concavified Power Utility functions (different risk aversion)
}
\examples{
PowerUtility_con_3(gamma1=2,gamma2=4,gamma3=5,b=1,a=1,K=1,x=1.6)
}
